requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
cachetools==4.2.4
colorama==0.4.6
decorator==4.4.2
Django==4.2
django-cors-headers==3.14.0
django-etc==1.4.0
django-hitcount==1.3.5
django-resized==1.0.2
django-taggit==3.1.0
django-tinymce==3.6.1
django-unicorn==0.50.0
djangorestframework==3.14.0
orjson==3.8.10
Pillow==9.5.0
pytz==2023.3
shortuuid==1.0.11
soupsieve==2.4.1
sqlparse==0.4.4
termcolor==2.2.0
termcolors==0.1.0
tzdata==2023.3
